pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "second-brain"
version = "0.1.0"
description = "A Second Brain powered by MCP, AI agents, and smart connectors."
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "PyGithub",
    "python-telegram-bot",
    "pyyaml",
    "python-frontmatter",
    "python-docx",
    "python-pptx",
    "openpyxl",
    "PyPDF2",
]

[project.scripts]
second-brain = "second_brain.server:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/second_brain"]

[tool.ruff]
src = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
